20161229 - add permission checks to group controller, remove unused ConcactsController, remove unused Person model, set paginate to 100 in RetreatsController

+++ app/Http/Controllers/GroupsController.php
@@ -22,14 +22,12 @@ class GroupsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
-    {
-        // TODO: figure out how to get parish information
+    public function index() {
+        $this->authorize('show-group');
         $groups = \montserrat\Group::whereIsActive(1)->orderBy('name')->with('members')->get();
         foreach ($groups as $group) {
             $group->count = $group->members()->count();
         }
-        //dd($groups);
         return view('groups.index',compact('groups'));   //
     }
 
@@ -38,9 +36,8 @@ class GroupsController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
-    {
-     
+    public function create() {
+        $this->authorize('create-group');
         return view('groups.create'); 
     
     }
@@ -51,8 +48,8 @@ class GroupsController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Request $request)
-    {
+    public function store(Request $request) {
+        $this->authorize('create-group');
         $this->validate($request, [
             'name' => 'required',
             'title' => 'required',
@@ -81,14 +78,11 @@ class GroupsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function show($id)
-    {
-        //
+    public function show($id) {
+       $this->authorize('show-group');
        $group = \montserrat\Group::findOrFail($id);
        $members = \montserrat\Contact::whereHas('groups', function($query) use ($id) {
             $query->whereGroupId($id)->whereStatus('Added');})->orderby('sort_name')->get();
-         
-       //dd($group);
        return view('groups.show',compact('group','members'));//
     
     }
@@ -99,13 +93,10 @@ class GroupsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($id)
-    {
-        //
+    public function edit($id) {
+        $this->authorize('update-group');
         $group = \montserrat\Group::findOrFail($id);
-        
         return view('groups.edit',compact('group'));
-    
     }
 
     /**
@@ -115,8 +106,8 @@ class GroupsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
-    {
+    public function update(Request $request, $id) {
+        $this->authorize('update-group');
         $this->validate($request, [
             'name' => 'required',
             'title' => 'required',
@@ -135,7 +126,7 @@ class GroupsController extends Controller
        
         $group->save();
     
-        return Redirect::action('ContactsController@index');//
+        return Redirect::action('PersonsController@index');//
     }
 
     /**
@@ -144,12 +135,9 @@ class GroupsController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
-    {
-        //
+    public function destroy($id) {
+        $this->authorize('delete-group');
         \montserrat\Group::destroy($id);
         return Redirect::action('GroupsController@index');
-    
     }
-    
 }
